+++ cw-orch-interchain/tests/common/mod.rs
@@ -1,2 +1,1 @@
-pub mod bank;
 pub mod ica_demo;