+++ cw-orch/examples/complex_testnet_daemon.rs
@@ -98,14 +98,11 @@ pub fn main() {
         .unwrap();
     // We send some funds to the counter contract
     let contract_addr = counter.address().unwrap();
+
     daemon
-        .rt_handle
-        .block_on(
-            daemon
-                .sender()
-                .bank_send(&contract_addr, coins(50_000, denom.clone())),
-        )
+        .bank_send(&contract_addr, &coins(50_000, denom.clone()))
         .unwrap();
+
     // We verify they have received their funds
     assert_eq!(
         daemon